+++ imports/plugins/core/ui/client/components/calendarPicker/calendarPicker.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 import omit from "lodash/omit";
 import { registerComponent } from "@reactioncommerce/reaction-components";
-import DayPickerRangeController from "./dynamicCalendarPicker";
+import DayPickerRangeController from "./dayPickerRangeController";
 
 // CalendarPicker is a wrapper around react-dates DayPickerRangeController. Anything that works in react-dates should
 // work in CalendarPicker react-dates docs are available at: https://github.com/airbnb/react-dates
