+++ imports/plugins/core/ui/client/components/calendarPicker/dayPickerRangeController.js
